requirements.txt
django==1.11.29
Pillow==6.2.2
django-crispy-forms==1.9.0
djangorestframework==3.9.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
